Initial porting of loop summarization code Port loop summarization code to latest cgc-fuzzball. To make it works correctly with cb-multios scripts, we also implemented -concolic-read and flush stderr.
In DARPA CGC -concolic-receive was used to perform concolic execution with given inputs.
To work with general Linux version of those CBs, a similar feature for read() instead of receive() is required.

Since the cgc-branch FuzzBALL use eprintf to print logs, we must flush stderr if we want to see outputs immediately.
It also make sense to flush stdout, since the outputs of executed binaries are usually redirected to stdout