Refactoring for context-ware loop summarization Refactor the majority of loopsum code to support context-aware loopsum.
When a function that contains a loop is called for multiple times, the status and IV values can be different.
To handle that correctly, we either regenerate loopsums for every occurence, or we generate a universe loopsum and plugin different values for each occurence.

The IV addr should be the operand part of the CODE rather than a value
And the code should be eval to a value at use(e.g. check exp)
More specific, the following changes has been made:
- replace iv addr with the offset from the base of stack frame
- change D_0 to the code exp of jmp cond and redesign GT data structure
- move the cond split part of eval_cjmp to loop_sum
- move check_loopsum to loop_record class
- check and apply loopsum when entering a loop, rather than at the first cjmp
- move run_sl ahead check_loopsum to fix DT inconsistency caused by last change
- refactor add_g to compute_distance, compute_dd, compute_ec
- remove iof_cache, since D0 can be different in different context (e.g. cond = {n == 100?}, n = 1 vs. n = 200
- remove code that apply loopsum in the same path from which FuzzBALL just create a loopsum
- refactor compute_lss to compute_precond and more
- remove is_gt_cond, clean_ivt/gt, renamve renew_ivt to update_ivt
- Update comments

The code now compiles again and works mostly correct on example/input-dependent, though context-awareness has yet to be test.
A visualized DT for input-dependent is included.